Add support for ScrambleSuit IAT obfuscation (Fixes #8)
obfsclient generates a random distribution of delays in 100 usec increments ranging from 0 to 10 msec.  I do not beleive that the fixed increment is fingerprintable since the delay is a delta on when tor actually sends a packet, and the timing of the former is not externally visible.

My implementation of IAT obfsucation is different from phw's in that, padding is only inserted when the frame being sent will fully drain the queued IAT TX buffer, and the code will combine writes up to the MTU.

phw's implementation:
1. Receive data from local tor
   1. Break up the received data into frames
   2. Add padding
   3. Schedule a timer
2. When the timer fires, dispatch a generated frame

My implementation:
1. Receive data from local tor
   1. Schedule a timer
2. When the timer fires
   1. Send up to one full sized frame of buffered data
   2. If the frame sent fully drains the buffered data, add padding
   3. If there is buffered data remaining, reschedule the timer