Ignore deprecation warnings if the host OS is Darwin (Fixes #30)
OSX deprecated OpenSSL due to ABI incompatibility issues, so the build breaks.

The right thing to do is pointed out in the bug, but this is also what tor
does, and more importantly my lone OSX box is close to a decade old, in a
cardboard box in the closet, and in a disassembled state, so the number of
active developers that need the warnings is 0.