Fix an assertion that occured if the outgoing connect call failed immediately
This was primarily visibile if the outgoing network interface was down
and was due to sending a socks response right as outgoing_connect()
returned, and then the bufferevent callback being called.

Code was changed to use the bufferevent callback to send a response as
long as the bufferevent was created successfully.