Apply pushback between incoming_/outgoing_ (Fixes #3)
The method chosen is to impose a soft limit on the write buffer of
each bufferevent and toggle EV_READ on the opposite bufferevent if
the limit is exceeded.  It is a "soft" limit because the write that
pushes the buffer size over the threshold can be any size.  The
current threshold is set to 64 KiB, and reads will be re-enabled
once the write buffer shrinks to 32 KiB or below.

It is possible that the thresholds are overly conservative, but
artificially lower performance is massively superior to potentially
buffering a huge amount of data.