Fix a potential bug found by clang's scan-build.
With all current uses of Ctr, the conditional I was missing is never
triggered since the paramenters are guaranteed to be sane, but this is
not guaranteed to be true in the future.

scan-build also points out a bunch of things in easylogging++ and gtest
but they don't look to be real problems.